Prem

+++ app/src/main/java/com/example/buysell2/Do/UserDo.java
@@ -6,12 +6,12 @@ import java.util.ArrayList;
 @SuppressWarnings("serial")
 public class UserDo implements Serializable {
     public ArrayList<SupplierMasterDo> listSupplierMaster = new ArrayList<>();
-    public int UP_ID = 0;
+    public long UP_ID = 0;
     public String UP_Name = "";
     public String UP_User_Type = "";
     public String UP_Email = "";
-    public int UP_Mobile_No = 0;
-    public int UP_UserID = 0;
+    public long UP_Mobile_No = 0;
+    public long UP_UserID = 0;
     public String UP_Password = "";
     public String UP_Status = "";
 }
